lib: rust_support: Remove heap allocation from logging
CString::new previously did a heap allocation to add a
null-terminator before passing the log message to C.
Allocations might not be allowed in all contexts where the
kernel may want to log so this changes fputs to fwrite to
avoid the need for a null-terminator and uses format_args!
to avoid allocation due to rust formatting internals.

Bug: None
Change-Id: Ibbddb64ff003618694983fb2ebdc3d5f78156b18
Signed-off-by: Ayrton Munoz <ayrton@google.com>