arch: arm64: Use a non-argument register in EL switch code
The entry code that switches the current exception level
to EL1 on entry uses the x0 register. This overwrites
the incoming x0 value from the boot protocol if the
exception level is not already EL1. Switch the register
to an unused one (x4 for now) to preserve the original
value of x0.

Bug: 407895490
Test: Boot generic-arm64-virt-test-debug as BL33
Change-Id: Ia19af88d54b1cb6e56d2605eaa56115da3a2a4c4
Signed-off-by: Andrei Homescu <ahomescu@xwf.google.com>