dev: virtio: vsock: Add Rust bindings for interrupt handlers
We need these bindings in order to register interrupt handlers from
Rust. Specifically, this will allow us to switch the vsock-tipc bridge
from polling to an interrupt-driven mode of operation.

Note that we specify the CFI encoding for the `handler_return` type in
order to avoid a cross-language CFI check failure. Since the attribute
goes on a type that is bindgen'ed, we have tell bindgen to add the
`cfi_encoding` attribute via `MODULE_BINDGEN_FLAGS`. An alternative
solution would be to mark the Rust enum as `repr(C)` but that is not
currently supported by bindgen.

Bug: 390490408
Link: https://github.com/rust-lang/rust-bindgen/issues/3263
Test: build.py generic-arm64-virt-test-debug --skip-tests
Change-Id: I3b47a90cf676cc9e7345fd79a3c3982cde90c753
Signed-off-by: Per Larsen <perlarsen@google.com>