dev: virtio: vsock: Limit port zero to test builds
Only make port zero a special port where the first package contains the
tipc port name to connect to. This frees up port one to be used in the
port map but for now, we keep an entry in the port map to preserve the
port numbering.

Moreover, only enable port zero for test builds. Release builds should
use ports in the port map instead.

Bug: None
Test: boot guest VM and test that port zero is only available on test
  builds.
Change-Id: Ice797d8a50e40fb93cf343d1f0e8b0fa307199eb
Signed-off-by: Per Larsen <perlarsen@google.com>