dev: virtio: vsock-rust: Register device for VM destroy notifications
This commit uses the VM destruction notification to clean up TZ
resources for virtio-msg vsock devices. Once Trusty receives a VM
destruction notification it uses the client ID to find the
virtio-msg vsock device and signals the rx/tx loops and the
memory unmapping thread to stop. Once the rx/tx loops stop the
VsockDevice Drop impl signals an event which the virtio-msg unmap
memory thread waits for. At that point it resets the
`VirtioMsgDevice` for the next client.

Bug: 433491122
Change-Id: I59efb546e2fea31cb7cdd3c5cfa1dbdc3e6ad3e6
Signed-off-by: Ayrton Munoz <ayrton@google.com>