dev: virtio: vsock-rust: Fix event queueing in vsock_rx_loop
When the `vsock_rx_loop` is in the `TipcConnecting` state it
may receive and queue up multiple events. Previously the
queueing behavior was incorrect because the vsock events
were pushed and popped from the end of a `Vec` so requeued
events would be handled in reverse order that they were
received. This fixes that behavior by replacing this `Vec`
with a `VecDeque` and pushing to the front rather than to
the back of the queue.

Bug: None
Change-Id: If6b8db36957ee39dcf2ab63a8345a12799bd3050
Signed-off-by: Ayrton Munoz <ayrton@google.com>