dev: virtio: vsock-rust: Send status byte on connection to special ports
When a TIPC port name is expected in the first message, the tx
loop needs to send a status byte to notify the client when the
TIPC connection is ready. In this case we should also not queue
any data received over vsock so this commits reverts that to the
behavior before ag/34296764.

Bug: 439981769
Change-Id: I3cfd29aa5baa8f04efe2911d16b6971b7e75552c
Signed-off-by: Ayrton Munoz <ayrton@google.com>