dev: virtio: vsock-rust: Initialize virtio-msg vsock devices
This commit initializes each vsock device found on the virtio-msg-ffa
bus by Trusty pVM builds. For each device found the virtio-msg init
hook calls `vsock_init` which creates threads for the vsock tx and rx
loops. For the tx loops it passes in `TransportKind::DriverFFAMsg`
to create the TIPC ports for the vsock-tipc bridge. It removes the
allow(dead_code) attribute in the vsock msg module.

Bug: 379677575
Change-Id: I4c51ef808a545e2a2ed3f6d50f077a1718a4bf97
Signed-off-by: Ayrton Munoz <ayrton@google.com>