lib: rust_support: handle: Add HandleCookie convenience wrapper
This commit adds HandleCookie as a newtype for a boxed T which is
expected to be used as a handle_ref cookie and requires a stable
address as a result. The definition uses UnsafeCell to opt-out of
Box asserting uniqueness over its contents (it may alias with
pointers which are piped through C code though never accessed
there) and to provide a simple way to get a pointer to the
contents. The arguments to `HandleRef::cookie` and `set_cookie`
were left as `*mut T` to allow the use of other kinds of cookies
(e.g. Arc<T>).

Bug: 433491122
Change-Id: Ib94f20dc2f09cd5a2ec8a5ca777bd6210dcf7c95
Signed-off-by: Ayrton Munoz <ayrton@google.com>