dev: virtio: vsock-rust: Remove c_void pointers in copy_nonoverlapping
copy_nonoverlapping uses the pointer type to determine how many
bytes to copy so this commit changes c_void pointers to u8 pointers
when it's called by the PCI vsock driver. c_void worked as
expected because the rust compiler internally represents c_void
as a `repr(u8)` enum but this is an implementation detail which
we should not rely on.

Bug: None
Change-Id: I15d5645543907d18ba1659eb8d948b1c980f836c
Signed-off-by: Ayrton Munoz <ayrton@google.com>