dev: virtio: vsock-rust: Only build msg module on aarch64
This crate's msg module assumes that usize == u64 in various places to
simplify the use of the existing virtio_drivers_and_devices Hal trait
with bus addresses rather than physical addresses. It also uses
various functions from the arm_ffa module which are only defined on
ARM64. These cause build errors on ARM32 so this commit ensures the
msg module is only built on ARM64. This allows including the
vsock-rust crate unconditionally for both ARM32 and ARM64 while making
sure that only the latter links in code for the virtio-msg transport.

Bug: 379677575
Change-Id: Ic53b9e9a6bf480c880db9588766c078eb55f02ce
Signed-off-by: Ayrton Munoz <ayrton@google.com>