make: Add alloc and trusty-std as default dependencies in the kernel
This commit adds alloc and trusty-std as default dependencies in
the kernel to ensure they are built before any other crates. The
latter does not include libstd, only the `#[global_allocator]`
needed for liballoc. This is a workaround for library.mk creating
spurious circular dependencies with rust in the kernel (b/450652789)
Once library.mk this is fixed it may be reverted to support
Trusty builds that don't make allocations in rust.

Bug: 450583155
Change-Id: Ie433c0f1038ffb6f290d3aed690881e7e2133881
Signed-off-by: Ayrton Munoz <ayrton@google.com>