feat(vsock): Add port for GateKeeper with THAL
This change reserves a vsock port for the GateKeeper HAL
service running within the Trusty VM.

A new build configuration, `TRUSTY_VM_INCLUDE_GATEKEEPER_WITH_THAL`,
is added to enable the `gatekeeper_with_thal` feature flag in
the vsock-rust module.

When enabled, this reserves port 13 for the
`android.hardware.gatekeeper.IGateKeeper_with_thal` service,
allowing the host to communicate with it.

Bug: 433513029
Test: m
Change-Id: I4960cdaf3eb09d7d1c56faf8083e4b655209a12a
Signed-off-by: Alice Wang <aliceywang@google.com>