dev: virtio: vsock-rust: Add method to get MSG_GET_DEVICES response
This commit also adds a new method to get the payload in a
virtio_msg response struct for types that cannot implement
zerocopy's traits (due to having flexible array members at the
end of the payload types themselves). In this case the method
returns a copy of the fixed-sized part of the payload and a slice
to the variably-sized part in the original buffer.

Bug: 379677575
Change-Id: Ib22a221e17a747084f231722a7cd207da96faf37
Signed-off-by: Ayrton Munoz <ayrton@google.com>