dev: virtio: vsock-rust: Add BusGetDevices for new virtio-msg device enumeration
This commit also removes the old VIRTIO_MSG_FFA_DEACTIVATE from
the set of msg IDs which a virtio-msg device may expect since it
has the same ID as the new VIRTIO_MSG_BUS_GET_DEVICES request. It
leaves the device-side device enumeration handling for a later
commit which will update the driver side in parallel.

Bug: 379677575
Change-Id: Ib6d97d82be61e38b2a30b448de50b08d84cb52e2
Signed-off-by: Ayrton Munoz <ayrton@google.com>