dev: virtio: vsock-rust: Add methods to access buffer as virtio_msg struct
This commit adds methods to reinterpret a suitably-sized buffer
as a virtio_msg struct using the bindgen-generated types from the
latest version of the virtio-msg headers. Unlike the previous
version the new headers do not distinguish between bus and device
requests in the type system so there is no equivalent to the old
virtio_msg_ffa struct. This version also uses a flexible array
member to represent the virtio_msg struct payload. Bindgen
generates a ZST at this offset and provides unsafe methods to
access it as a `&[u8]` so the methods added in this commit only
provide safe access to the fixed-size header in virtio_msg
structs.

Bug: 379677575
Change-Id: Ie136fb8de834ebe0bc6fc81824e13abdf580a8af
Signed-off-by: Ayrton Munoz <ayrton@google.com>