dev: virtio: vsock-rust: Update FFA_BUS_MSG_AREA_SHARE to virtio-msg DEV2
This commit updates FFA_BUS_MSG_AREA_SHARE to match the DEV2
version of the virtio-msg spec. Even though the new spec allows
16-bit area IDs this commit leaves the limit of 255 memory
regions and requires that they have the upper 8 bits zeroed. The
limit on the number of memory regions is a consequence of the
device in Trusty only supporting direct messages and may be
either made configurable in a follow-up or removed once indirect
messages are supported.

Bug: 379677575
Change-Id: Icd743f297747dd73fe439a5bfdf7c3c7df5bf777
Signed-off-by: Ayrton Munoz <ayrton@google.com>