dev: virtio: vsock-rust: Check negotiated version before using FF-A 1.2 ABIs
This modifies the virtio-msg vsock driver to check the FF-A
version negotiated by the arm_ffa init hook before attempting to
use the FFA_MSG_SEND_DIRECT_REQ2 ABIs. In the case that a version
lower than 1.2 or higher than 2.0 has been negotiated the driver
logs a message and skips the rest of its initialization without
panicking.

Bug: 457317816
Change-Id: I129d766b71b7a26f50092ce994600c019e429452
Signed-off-by: Ayrton Munoz <ayrton@google.com>